refactor(benchmark): split manager.go session execution into runner.go
manager.go was 925 lines, over the 800-line ceiling. Move the benchmark
session-execution cluster (RunCapabilityDetection, runSingleBenchmark,
collectSingleBenchmarkOutput, runBenchmarks, runBenchmarkTask,
processBenchmarkOutput, finalizeBenchmarkSession, submitBatchIfReady, and the
killAndDrain helper + benchmarkBatchSize/processFlushTimeout consts) into a new
runner.go (431 lines), leaving manager.go (509 lines) with the public API, cache,
and submission logic. Pure move, behavior unchanged.

Resolves R16.
